2026/uddalak_multimodal_ai_agent_eval/backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    # Primary providers for this PoC (keys the user actually has)
    GEMINI_API_KEY: str = os.getenv("GEMINI_API_KEY", "")
    GROQ_API_KEY: str = os.getenv("GROQ_API_KEY", "")
    # Optional / legacy
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY", "")
    ANTHROPIC_API_KEY: str = os.getenv("ANTHROPIC_API_KEY", "")


settings = Settings()
