2026/uddalak_multimodal_ai_agent_eval/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.95.0",
    "fastapi>=0.135.3",
    "google-genai>=1.0",
    "groq>=0.28.0",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "openai>=2.31.0",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
